requirements.txt
selenium==4.11.0
python-dotenv==1.0.0